chore: rename fillmage.ts to fillImage.ts
