+++ packages/layers/src/point/models/fillImage.ts
